+++ lib/legion/extensions/node/actors/cluster_control.rb
@@ -1,5 +1,7 @@
 # frozen_string_literal: true
 
+require 'legion/extensions/node/control_auth'
+
 module Legion
   module Extensions
     module Node
@@ -27,6 +29,10 @@ module Legion
 
           def generate_task?
             false
+          end
+
+          def process_message(message, metadata, delivery_info)
+            Legion::Extensions::Node::ControlAuth.verify!(super)
           end
         end
       end