+++ lib/legion/extensions/node/transport/messages/cluster_settings.rb
@@ -1,5 +1,7 @@
 # frozen_string_literal: true
 
+require 'legion/extensions/node/control_auth'
+
 module Legion
   module Extensions
     module Node
@@ -23,11 +25,11 @@ module Legion
             end
 
             def message
-              {
+              Legion::Extensions::Node::ControlAuth.sign(
                 function: 'update_settings',
                 settings: @options[:settings],
                 restart:  @options.fetch(:restart, false)
-              }
+              )
             end
 
             def validate